Three changes:
CURRENTLY IMPORTING: a new field.
This field is used by PATCH^A1AECL1 to decide whether to populated when
we actually release a patch out of the door.

DERIVED FROM PATCH is now spelled correctly.

STATUS OF PATCH has more extensive Description an Technical Description
to document how it is now supposed to work in the new design for PM 2.4.